pyproject.toml
[project]
name = "rag-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.12"
dependencies = [
    "chromadb>=1.5.5",
    "langchain>=1.2.11",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.2.3",
]
